requirements.txt
torch
torchvision
transformers
sentence-transformers
easyocr
scikit-learn
fastapi
uvicorn
python-multipart
pillow
pyyaml
tqdm
pandas